+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueEvent.kt
@@ -11,3 +11,7 @@ data class CurrentFacilityLoaded(val facility: Facility) : OverdueEvent()
 data class CallPatientClicked(val patientUuid: UUID) : OverdueEvent() {
   override val analyticsName = "Overdue Screen:Call Patient clicked"
 }
+
+data class PatientNameClicked(val patientUuid: UUID) : OverdueEvent() {
+  override val analyticsName = "Overdue Screen:Patient name clicked"
+}