+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueScreen.kt
@@ -16,10 +16,14 @@ import org.simple.clinic.contactpatient.ContactPatientBottomSheet
 import org.simple.clinic.di.injector
 import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.router.screen.ScreenRouter
+import org.simple.clinic.summary.OpenIntention
+import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.util.UserClock
+import org.simple.clinic.util.UtcClock
 import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.widgets.PagingItemAdapter
 import org.simple.clinic.widgets.visibleOrGone
+import java.time.Instant
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import java.util.UUID
@@ -39,6 +43,9 @@ class OverdueScreen(
 
   @Inject
   lateinit var userClock: UserClock
+
+  @Inject
+  lateinit var utcClock: UtcClock
 
   @Inject
   @Named("full_date")
@@ -121,6 +128,14 @@ class OverdueScreen(
           overdueRecyclerView.visibleOrGone(isVisible = areOverdueAppointmentsAvailable)
         }
         .subscribe(overdueListAdapter::submitList)
+  }
+
+  override fun openPatientSummary(patientUuid: UUID) {
+    screenRouter.push(PatientSummaryScreenKey(
+        patientUuid = patientUuid,
+        intention = OpenIntention.ViewExistingPatient,
+        screenCreatedTimestamp = Instant.now(utcClock)
+    ))
   }
 
   interface Injector {